requirements.txt
Jinja2==3.1.2
python-dotenv==1.0.0
Requests==2.31.0
